Move as-sha256 to packages/
