+++ packages/as-sha256/assembly/index.ts
