+++ packages/as-sha256/examples/run-wasm.js
