+++ packages/as-sha256/src/index.js
