+++ packages/as-sha256/test/benchmark.test.js
