+++ packages/as-sha256/test/index.spec.js
